requirements.txt
openai>=1.40.0
python-dotenv
requests
beautifulsoup4
lxml
trafilatura